some more fixes and unit tests
