+++ test/ui/common/attributes/AttributeListTableActions.test.js
